+++ CHANGELOG.md
@@ -62,6 +62,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Adds format date to csv downloads
 - Adds CSV download for interpreting results column headers
 - Adds lat/lng coordinates on cursor move
+- Adds database selection
 
 ### Changed